requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-registration==3.1
django-rest-auth==0.9.5
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-webpack-loader==0.7.0
djangorestframework==3.11.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
whitenoise==5.1.0
zope.interface==5.1.0
